fix: avoid circular dependency
