+++ src/lib/bridge/handlers/fetch_steam_trades.ts
@@ -1,6 +1,5 @@
-import {getTradeOffersWithDescriptionFromAPI} from '../../alarms/trade_offer';
+import {ExtendedOfferStatus, getTradeOffersWithDescriptionFromAPI} from '../../alarms/trade_offer';
 import {rgDescription} from '../../types/steam';
-import {TradeOfferState} from '../../types/steam_constants';
 import {SimpleHandler} from './main';
 import {RequestType} from './types';
 
@@ -11,22 +10,6 @@ export interface FetchSteamTradesResponse {
     sent: ExtendedOfferStatus[];
     descriptions: rgDescription[];
     steam_id?: string | null;
-}
-
-export interface ExtendedOfferStatus {
-    offer_id: string;
-    state: TradeOfferState;
-    given_asset_ids?: ExtendedSingleOffer[];
-    received_asset_ids?: ExtendedSingleOffer[];
-    time_created?: number;
-    time_updated?: number;
-    other_steam_id64?: string;
-}
-
-export interface ExtendedSingleOffer {
-    assetid: string;
-    classid: string;
-    instanceid: string;
 }
 
 export const FetchSteamTrades = new SimpleHandler<FetchSteamTradesRequest, FetchSteamTradesResponse>(
